Fix Kea pool formatting and reorganize migrate tests
